Make markers appear at user's actual location
- Add logic to pass each user's human-readable location to
a geocoding API and retrieve the corresponding
latitude and longitude of the location. To reduce number
of API calls, an API call is not made if the location is
obviously invalid (ex. empty string.) For testing purposes
only a maximum of 100 API calls are made currently, since
there is a limit of 2500 calls per day with the test API key.

- Also, add logic to display markers at the user's actual
location in the world, since the correct latlng information
has been obtained. Only users with a valid location are
displayed on the map.

- Change the marker appearance to that of a circle.