remove unneeded code used by infeasible approach
Since we are limited to only 2500 requests/day for
the geocoding API in production, requests must be
kept at a minimum. Here's some of the facts:
- Everybody who has ever contributed to OneCommunity
  will be displayed on the map. This means a lot of
  different locations. A request for each unique
  location will contribute a lot towards this limit.
- A public facing version of this map is planned.
  and for every new user who views the map, there
  will be a ton of API calls made since their
  browser has not cached the data yet.
- The cache is implemented via localStorage, and
  if it is cleared, a bunch of API requests will
  be made again. This means somebody could -
  intentionally or unintentionally, make us reach
  our limit in the span of a few minutes.

Therefore, the current approach of caching the data
in the browser as a rate limiting technique is not
feasible given our siutation. Instead, a better way
is to store the latlng as a property of a user's
profile in the database and only call the API to
update the latlng when a new user is created or when
their location changes. Then, when someone visits
the map, there are zero API calls made since the
locations have been precomputed and stored in the
backend. Therefore:
- remove caching logic
- add mock data similar to how the new response
  from the backend would look like