refactor marker rendering logic to be cleaner
Refactor logic so that the explicit conditional
check to render or not render a marker based on
whether the location is valid or not is removed,
and instead the array of users is filtered
beforehand to only include users with valid
locations. Makes the JSX cleaner. Also, added
an example user with no valid location to show
how the map code handles it.