See User Management Tab (Full Functionality)
Owner can add “See User Management Tab” permission to a user that does
not inherently have that permission (i.e., a Volunteer). The volunteer
user, after logging out and logging back in to HGN app, will now see
the “Other Links” button with the “User Management” option available.
And the volunteer user is now able to visit the “/usermanagement”
protected route without being redirected to “/dashboard”, and is able
to create, delete, and update users.