Additional text on Dashboard Badges for personal best
This commit is for part c of 'Differentiate Personal Max Badge'. Added a function that checks if the latest badge earned is the personal best badge. In Dashboard badges section, added a section of text that will appear when the latest badge earned is personal best badge. The correct number of hours displayed here needs to be fixed in another task WIP.