add asign model
Fix the issue of the # 1673 that the page is not showing. Now all the components can be open correctly but:
(1) auto select teams and project field does not work. This cause failure to add new title
(2) asiign team function does not work
(2) Not sure how GDocs works in current code